requirement.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
py-mon==2.0.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
watchdog==4.0.2
yarl==1.9.4
